+++ tests/rules/exporter/test_rules2excel.py
@@ -1,4 +1,3 @@
-
 from cognite.neat.rules.exporter.rules2excel import RulesToExcel